requirements.txt
numpy==1.22.0
scikit-learn==1.2.2
scipy==1.10.1
torch==2.0.0
torchvision==0.15.0
tqdm==4.65.0
Pillow==9.4.0
pandas==2.0.1
pytorch-lightning==1.9.5
PyYAML==6.0
opencv-python==4.8.1.78
matplotlib==3.7.1